+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @embeddedchat/react
 
+## 0.1.8
+
+### Patch Changes
+
+- Fixed: Implement Token Handling Utilities Compatible with SSR
+  Fixed: Broken Behavior of Emoji picker in toolbar
+  Fixed: Message reactions size
+  Fixed: Remove unused menu options
+  Fixed: Fixed fullscreen action button
+- Updated dependencies
+  - @embeddedchat/api@0.0.2
+
 ## 0.1.5
 
 ### Patch Changes